# swagger.json
{
"swagger": "2.0",
"info": {
    "version": "1.0.0", 
    "title": "Video Game API",
    "description": "A Simple API to Store Video Game Data"
},
"host": "localhost:3000",
"basePath": "/",
"consumes": ["application/json"],
"produces": ["application/json"],
"paths": {
    "/allgames": {
        "get": {
            "tags": ["Paths"],
            "summary": "Lists All Games",
            "responses": {
                "200": {
                    "description": "OK",
                    "schema":{
                        "type": "object",
                        "example": [
                            { 
                              "_id": "641380184e69661ee7e5f0ea",
                              "": 507,
                              "name": "Chaos Field: New Order",
                              "platform": "Playstation 2",
                              "developers": "MileStone Inc.",
                              "publisher": "MileStone Inc.",
                              "link": "",
                              "releaseDate": "2005-12-15JP",
                              "JP": "Released",
                              "EU/PAL": "Unreleased",
                              "NA": "Unreleased"
                            },
                            {
                              "name": "Acid Drop",
                              "platform": "Atari",
                              "developers": "Dennis Kiss[11]",
                              "publisher": "Salu Ltd",
                              "releaseDate": "1992",
                              "genres": "Strategy",
                              "notes": "PAL-only; puzzle game; last game released. Atari 2600s stopped being manufactured"
                            },
                            {
                              "_id": "641380184e69661ee7e5e379",
                              "meta_score": "",
                              "name": "Super Mario 3D World + Bowser's Fury",
                              "platform": "Switch",
                              "releaseDate": "Feb 12, 2021",
                              "user_score": "",
                              "link": "/game/switch/super-mario-3d-world-+-bowsers-fury",
                              "esrb_rating": "",
                              "developers": "['Nintendo']",
                              "genres": "['Action', 'Platformer', '3D']"
                            }
                      ]
                    }
                }  
            }
        },
        "post": {
            "tags": ["Create"],
            "summary": "Creates New Entry to All Games",
            "consumes": "application/json",
            "parameters":[ 
                {
                 "in": "body",
                 "name": "body",
                 "schema": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "String"
                        },
                        "platform": {
                            "type": "String"
                        },
                        "developers": {
                            "type": "String"
                        },
                        "releaseDate": {
                            "type": "String"
                        },
                        "genres": {
                            "type": "String"
                        }
                    }
                 }
                }
            ],
            "responses": {
                "201": {
                    "description": "Created",
                    "schema": {
                        "type": "object",
                        "example": {
                            "name": "Super Smash Bros",
                            "platform": "Switch",
                            "releaseDate": "Oct 13, 2020",
                            "developers": "Nintendo",
                            "genres": "['Action', 'Fighting', '2D']"
                        }
                    }
                }
            }
        }
    },
...
    "/allgames/{id}": {
        "put": {
            "tags": ["Update"],
            "summary": "Update Existing Entry for All Games",
            "consumes": "application/json",
            "parameters":[ 
                {
                 "in": "path",
                 "required": true,
                 "schema": {
                    "type": "interger",
                    "minimum": 1
                 }
                },
                {
                 "in": "body",
                 "required": true,
                 "name": "body",
                 "schema": {
                    "type": "object",
                    "properties": {
                        "name": {
                            "type": "String"
                        },
                        "platform": {
                            "type": "String"
                        },
                        "developers": {
                            "type": "String"
                        },
                        "releaseDate": {
                            "type": "String"
                        },
                        "genres": {
                            "type": "String"
                        }
                    }
                 }
                }
            ],
            "responses": {
                "200": {
                    "description": "OK",
                    "schema": {
                        "type": "object",
                        "example": {
                            "name": "Super Smash Bros",
                            "platform": "Switch",
                            "releaseDate": "Oct 13, 2020",
                            "developers": "Nintendo",
                            "genres": "['Action', 'Fighting', '2D']"
                        }
                    }
                }
            }
        },
        "delete": {
            "tags": ["Delete"],
            "summary": "Deletes Entry to All Games",
            "consumes": "application/json",
            "parameters":[ 
                {
                 "in": "path",

                 "required": true,
                 "schema": {
                    "type": "interger",
                    "minimum": 1
                 },
                 "description": "Game '_id'"
                }
            ],
            "responses": {
                "202": {
                    "description": "Accepted",
                    "schema": {
                        "type": "object",
                        "example": {
                            "name": "Super Smash Bros",
                            "platform": "Switch",
                            "releaseDate": "Oct 13, 2020",
                            "developers": "Nintendo",
                            "genres": "['Action', 'Fighting', '2D']"
                        }
                    }
                }
            }
        }
    },
...
  },
...
}